Updated Branche so its up to date (#2)
* Create node.js.yml

* Added test script

* Update node.js.yml

* Added package Lock for auto testing

* Update .gitignore

* Updated YT Downloader

* Implemtened Safty Features

* Added some files for tomorrow

* Stared reworking backend

* Updated User Login

The Way a User is stored is now different.
Every User does now have a uuid and is its own Objets

* You can now add Users from the admin page

* Bug Fixing

* Bug Fixing

* Fixed Typo

* 🐛Fixes

* Started exporting routes
