--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-        <title>Chromecast</title>
-</head>
-<body>
-        <form>
-                <button type="button" id="cast">Click To Cast</button>
-        </form>
-        <script type="text/javascript" src="script.js"></script>
-</body>
-</html>
