--- login.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-   <head>
-     <title>Loginformular</title>
-     <link rel="stylesheet" href="/style/login.css" type="text/css"></link>
-    </head>
-<body>
-<div class="container">
-  <h1>Login</h1>
-    <input type="text" id="username" name="user" placeholder="Benutzername">
-    <input type="password" id="password" name="password" placeholder="Passwort"><br>
-    <button type="submit" name="submit" id="submit" class="button">Login</button>
-    <div class="wrongPass" id="wrongPass">
-        <b id="wrongPassText">Das Passwort oder der Benutzername ist falsch!</b>
-    </div>
-  <script type="text/javascript" src="/js/login.js"></script>
-</div>
-
-</body>
-</html>
