+++ package.json
@@ -10,7 +10,8 @@
   },
   "devDependencies": {},
   "scripts": {
-    "run": "node index.js"
+    "test": "node index.js test",
+    "debug" : "node index.js debug"
   },
   "repository": {
     "type": "git",
