+++ private/html/admin.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link rel="stylesheet" type="text/css" href="/private/style/admin.css">
+        <title>Settings</title>
+    </head>
+    <body>
+        <div id="container"></div>
+        <label for="newUser" id="FormLabel">Add new User</label>
+        <div id="newUser">
+            <input type="text" placeholder="Username" id="Name"></input>
+            <input type="password" placeholder="Password" id="Pass">
+            <input type="password" placeholder="Confirm Password" id="ConPass">
+            <label for="perm" id="permLabel">Permission:</label>
+            <select id="perm">
+                <option value="Admin">Admin</option>
+                <option value="User" selected>User</option>
+            </select>
+            <input type="button" value="Add User" id="submit">
+        </div>
+        <script type="text/javascript" src="/private/js/admin.js"></script>
+    </body>
+</html>
