+++ private/html/logs.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Logs</title>
+        <link rel="stylesheet" type="text/css" href="/private/style/logs.css">
+    </head>
+    <body>
+        <div id="holder">
+            <div class="item">
+                <button id="clear">
+                    Clear Console
+                </button>
+            </div>
+        </div>
+        <script type="text/javascript" src="/private/javascript/logs.js"></script>
+    </body>
+</html>
