+++ private/html/videoShow.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link rel="stylesheet" type="text/css" href="/private/style/videoShow.css">
+    </head>
+    <body>
+        <div id="container">
+        </div>
+    </body>
+    <script type="text/javascript" src="/private/js/videoShow.js"></script>
+</html>
