+++ private/style/admin.css
@@ -1,0 +1,100 @@
+* {
+    background-color: darkslategray;
+}
+
+
+label {
+    font-weight: bold;
+    display: block;
+    margin: 5px;
+    margin-bottom: 0px;
+    margin-top: 0px;
+    margin-left: 0px;
+    width: 30px;
+}
+
+.User {
+    margin: 10px;
+}
+
+/* The switch - the box around the slider */
+.switch {
+    position: relative;
+    display: inline-block;
+    width: 40px;
+    height: 20px;
+  }
+  
+  /* Hide default HTML checkbox */
+  .switch input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+  }
+  
+  /* The slider */
+  .slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: red;
+    -webkit-transition: .4s;
+    transition: .4s;
+  }
+  
+  .slider:before {
+    position: absolute;
+    content: "";
+    height: 10px;
+    width: 10px;
+    left: 4px;
+    bottom: 4px;
+    background-color: blue;
+    -webkit-transition: .4s;
+    transition: .4s;
+  }
+  
+  input:checked + .slider {
+    background-color: greenyellow;
+  }
+  
+  input:checked + .slider:before {
+    -webkit-transform: translateX(200%);
+    -ms-transform: translateX(200%);
+    transform: translateX(200%);
+  }
+  
+  /* Rounded sliders */
+  .slider.round {
+    border-radius: 34px;
+  }
+  
+  .slider.round:before {
+    border-radius: 50%;
+  }
+
+#newUser input {
+  border: black;
+  border-style: inset;
+  border-width: 1px;
+}
+
+
+#permLabel {
+  display: inline;
+  font-weight: bolder;
+}
+
+#FormLabel {
+  display: inline;
+  margin: 10px;
+}
+
+#newUser {
+  margin: 10px;
+  margin-top: 0;
+  display: block;
+}
