+++ public/html/notFound.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <script type="text/javascript" src="/public/js/notFound.js"></script>
+        <title>404 Not Found</title>
+    </head>
+    <body style="background-color: darkslategray;">
+        <b style="color: white; position: absolute; text-align:center; left: 50%; top: 50%; transform: translateX(-50%) translateY(-50%); font-size: 50px;">
+            The site you are trying to enter does not exist!
+        </b>
+    </body>
+</html>
