--- public/style/logs.css
@@ -1,35 +1,0 @@
-* {
-    background-color: darkslategray;
-}
-
-div.item {
-    margin: 5px;
-    padding: 2px;
-}
-
-label {
-    color: deepskyblue;
-    font-weight: bold;
-    display: block;
-    margin: 5px;
-    width: 30px;
-    padding: 2px;
-}
-
-button {
-    background: gray;
-    border-width: 1px;
-    border-color: black;
-}
-
-div.logItem {
-    margin-left: 5px;
-    margin-top: 2px;
-    margin-bottom: 2px;
-}
-
-b {
-    padding: 2px;
-    background: gray;
-    margin: 3px;
-}
