+++ routes/index.js
@@ -1,0 +1,6 @@
+var express = require('express');
+var router = express.Router();
+
+router.use('/', require('./backend/index'));
+
+module.exports = router;
