--- videoPlayer/admin.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" type="text/css" href="/style/admin.css">
-        <title>Settings</title>
-    </head>
-    <body>
-        <div id="container">
-            
-        </div>
-
-        <script type="text/javascript" src="/js/admin.js"></script>
-    </body>
-</html>
