--- videoPlayer/logs.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Logs</title>
-        <link rel="stylesheet" type="text/css" href="/style/logs.css">
-    </head>
-    <body>
-        <div id="holder">
-            <div class="item">
-                <button id="clear">
-                    Clear Console
-                </button>
-            </div>
-        </div>
-        <script type="text/javascript" src="/js/logs.js"></script>
-    </body>
-</html>
