--- videoPlayer/videoShow.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" type="text/css" href="/style/videoShow.css">
-    </head>
-    <body>
-        <div id="container">
-        </div>
-    </body>
-    <script type="text/javascript" src="/js/videoShow.js"></script>
-</html>
